Make tests independent from default branch
The default branch name when not explicitely specified,
comes from the init.defaultBranchName setting, which is
`master` by default.

Remove this environment dependency by setting it explicitely,
should someone have it defined to something different from `master`.

Refactored `initRepo` to be used in the tests initialization
to avoid code repetition.

Bug: Issue 15891
Change-Id: Ib8fe1aa59b7250c94605a4160a681018f4c1cc5c