Remove dependency from parallel-collections to avoid CVE-2022-36944
The parallel-collections 1.0.4 depends on the scala-library:2.13.6
which is subject to CVE-2022-36944, causing the whole of gatling-git
to be flagged as insecure because of its dependencies.
The advantages of creating a random commit in parallel are really
minor and not worth the flagging as insecure library.

Remove the parallelism and the dependency on parallel-collections
and therefore clear the library from being flagged as insecure.

Change-Id: I530a2457dee03e6f216e48f1045fc7b1a7c93976