Remove Dockerfile
Going forward we'd like to publish this artifact as a regular JAR
instead of a fat JAR so it's packaged as a library to be used in other
projects that can define their own Dockerfile.

In order to achieve this we've removed the Dockerfile in this project so
we don't need the fat JAR.

If we feel strongly about including a Dockerfile in this repository, we
could alternatively add another sbt project that provides a fat JAR for
the Dockerfile to consume, allowing us to keep publishing the regular
JAR and only using the fat JAR locally for the Docker container.

This change removes the Dockerfile, a subsequent change will publish the
regular JAR.

Change-Id: I03836fed45aeaabd9b707565481c4671e90dd8c4