Allow to delete workdir after clone
When running the Clone command for load-testing purposes
the client workdir is automatically generated using the
System.nanoTime() for assuring that parallel requests do
not clash between each other.

For longer simulations, the number of cloned workspaces
can be overwhelming and fill up the local filesystem very
quickly.

Introduce a new option deleteWorkdirOnExit on the Clone
operation for allowing a simulation to remove the cloned
workspace and preventing potential out of disk space
errors.

Bug: Closes #3
Change-Id: Ia9bf6e189e5e4febeffbd332ea7f69900712bbd1