Propagate work directory to the Fetch request
The Fetch request was not previously made aware of the
local repository working directory override and
therefore was causing the creation of a new
empty directory from scratch at every request
execution.

From a server's perspective, any fetch was considered
as it was a full clone because the local directory
always started as an empty local Git repository.

Change-Id: I46664c2734e70397e8a21436ca99d411558775b8